Created user context and auth context
User and auth context created. These include /dashboard being wrapped by the authProvider. This means that on every refresh inside the /dashboard we are checking to see if the user's token allows them to access this site. If not they will be sent back to signup/login page or we can send them to home